bot/requirements.txt
line-bot-sdk
Flask
mysql-connector-python
configparser
requests
